xom: Allow re-running do_compile
do_compile modified the source directory destructively (mv and rm),
avoiding its re-run.

Fixes:
ERROR: xom-1.2.10-r0 do_compile: Function failed: do_compile (log file is located at /workdir/build/tmp/work/dbfp4-poky-linux/xom/1.2.10-r0/temp/log.do_compile.1523)
ERROR: Logfile of failure stored in: /workdir/build/tmp/work/dbfp4-poky-linux/xom/1.2.10-r0/temp/log.do_compile.1523
Log data follows:
| DEBUG: Executing shell function do_compile
| mv: cannot stat '/workdir/build/tmp/work/dbfp4-poky-linux/xom/1.2.10-r0/git/jaxen': No such file or directory
| WARNING: exit code 1 from a shell command.
| ERROR: Function failed: do_compile (log file is located at /workdir/build/tmp/work/dbfp4-poky-linux/xom/1.2.10-r0/temp/log.do_compile.1523)
NOTE: recipe xom-1.2.10-r0: task do_compile: Failed
ERROR: Task (/workdir/repo/poky/../meta-java/recipes-core/xml-commons/xom_1.2.10.bb:do_compile) failed with exit code '1'
ERROR: mc:inteli7:xom-1.2.10-r0 do_compile: Function failed: do_compile (log file is located at /workdir/build/tmp/work/corei7-64-poky-linux/xom/1.2.10-r0/temp/log.do_compile.1522)

+++ recipes-core/xml-commons/xom_1.2.10.bb
@@ -28,14 +28,22 @@ inherit java-library
 
 DEPENDS = "fastjar-native virtual/javac-native xerces-j xalan-j"
 
+do_move_jaxen() {
+    if [ -d ${WORKDIR}/git ] ; then
+        mv ${WORKDIR}/git/jaxen ${WORKDIR}/jaxen-${PV_jaxen}
+        rm -rf ${WORKDIR}/git
+    fi
+
+}
+
+addtask move_jaxen after do_removebinaries before do_patch
+
 do_compile() {
     mkdir -p build
 
     oe_makeclasspath cp -s xercesImpl xalan2
     cp=build:$cp
 
-    mv ${WORKDIR}/git/jaxen ${WORKDIR}/jaxen-${PV_jaxen}
-    rm -rf ${WORKDIR}/git
     scp="${WORKDIR}/jaxen-${PV_jaxen}/src/java/main"
 
     javac -sourcepath src:$scp -cp $cp -d build `find src -name "*.java" -and -not \( -wholename "*tests*" -or -wholename "*samples*" -or -wholename "*tools*" \)`
